# express-back/docs/swagger.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "哈尔滨学院校园综合服务平台 API",
    "version": "1.0.0",
    "description": "校园综合服务平台后端接口文档。"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Current server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/admin/audit": {
      "get": {
        "tags": [
          "Admin Audit"
        ],
        "summary": "审计日志管理路由",
        "operationId": "AuditLogController_getAuditLogs",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Admin Audit"
        ],
        "summary": "Delete Audit Logs",
        "operationId": "AuditLogController_deleteAuditLogs",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/audit/stats": {
      "get": {
        "tags": [
          "Admin Audit"
        ],
        "summary": "Get Audit Log Stats",
        "operationId": "AuditLogController_getAuditLogStats",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/audit/operators": {
      "get": {
        "tags": [
          "Admin Audit"
        ],
        "summary": "Get Operators",
        "operationId": "AuditLogController_getOperators",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/audit/high-risk": {
      "get": {
        "tags": [
          "Admin Audit"
        ],
        "summary": "Get High Risk Logs",
        "operationId": "AuditLogController_getHighRiskLogs",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/audit/{id}": {
      "get": {
        "tags": [
          "Admin Audit"
        ],
        "summary": "Get Audit Log Detail",
        "operationId": "AuditLogController_getAuditLogDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/audit/delete": {
      "post": {
        "tags": [
          "Admin Audit"
        ],
        "summary": "Delete Audit Logs",
        "operationId": "AuditLogController_deleteAuditLogs",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}